requirements.txt
streamlit
geemap
streamlit_folium
google-auth
geopy
flask
